converted app, index, server to typescript; configured tsconfig, modified webpack
