--- client/index.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { App } from './App.jsx';
-
-const container = document.getElementById('root');
-const root = createRoot(container); // createRoot(container!) if you use TypeScript
-
-root.render(
-  <App />
-);
