+++ package.json
@@ -3,11 +3,10 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
-  "type": "module",
   "scripts": {
-    "start": "node server/server.js",
-    "build": "webpack",
-    "dev": "concurrently \"webpack-dev-server --open --hot \" \"nodemon ./server/server.js \""
+    "start": "node --loader ts-node/esm ./src/server/server.ts",
+    "build": "tsc",
+    "dev": "concurrently \"webpack-dev-server --open --hot \" \"node --loader ts-node/esm ./src/server/server.ts \""
   },
   "author": "Judy Wu",
   "license": "ISC",
@@ -26,11 +25,18 @@
     "@babel/core": "^7.21.0",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
+    "@types/html-webpack-plugin": "^3.2.6",
+    "@types/node": "^18.14.6",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "babel-loader": "^9.1.2",
     "concurrently": "^7.6.0",
     "html-webpack-plugin": "^5.5.0",
     "nodemon": "^2.0.20",
     "path": "^0.12.7",
+    "ts-loader": "^9.4.2",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1"
